Начал выполнение задания 8.1
