+++ js/mocks/data.js
@@ -1,5 +1,6 @@
 import {getDecimalRandom, getRandom, getRandomArrayElement, getRandomArray} from './util.js';
-import {AVATAR_COUNT, TYPE_HOUSE, CHECKINS, CHECKOUTS, FEATURES, PHOTOS} from './vars.js';
+import {AVATAR_COUNT, TYPE_HOUSE, CHECKINS, CHECKOUTS, PHOTOS} from './vars.js';
+import {FEATURES} from '../vars.js';
 
 /**
  *  Создает объект
@@ -14,7 +15,7 @@ const generateOffer = () => {
       avatar: `img/avatars/user0${getRandom(1, AVATAR_COUNT)}.png`,
     },
     offer: {
-      title: undefined, //'Сдача недвижимости',
+      title: 'Сдача недвижимости',
       address: `${locationX}, ${locationY}`,
       price: getRandom(0, 100000),
       type: getRandomArrayElement(TYPE_HOUSE),
